Added the gnu-efi library and finished my UEFI pathway

+++ Makefile
@@ -1,11 +1,11 @@
 disk=./build/diskimage.dd
 parted=parted -s $(disk) unit s
 CC=x86_64-w64-mingw32-gcc
-CFLAGS=-Ipath/to/gnu-efi/inc -Ipath/to/gnu-efi/inc/x86_64 -Ipath/to/gnu-efi/inc/protocol
-qemu=qemu-system-x86_64 -hda $(disk)
-LDFLAGS=--nostdlib -Wl,-dll -shared -Wl, --subsystem,10 -e efi_main
+CFLAGS=-ffreestanding -Ignu-efi/inc -Ignu-efi/inc/x86_64 -Ignu-efi/inc/protocol -Ignu-efi/lib -c
+qemu=qemu-system-x86_64
+LDFLAGS=-nostdlib -Wl,-dll -shared -Wl,--subsystem,10 -e efi_main
 
-build:
+build:clean
 	mkdir -p build
 	nasm -fbin ./src/bootloader.asm -o ./build/bootloader.bin
 	dd if=/dev/zero of=$(disk) bs=1M count=10
@@ -13,13 +13,13 @@ build:
 	$(parted) mkpart primary 34 100
 	dd if=./build/bootloader.bin of=$(disk) bs=512 count=1 conv=notrunc
 
-clean:build
-	rm -r build
+clean:
+	if test -d build; then rm -r build; fi
 
-run:$(disk)
-	$(qemu)
+run:build
+	$(qemu) -hda $(disk)
 
-buildUEFI:
+buildUEFI:clean
 	mkdir -p build
 	nasm -fbin ./src/bootloader.asm -o ./build/bootloader.bin
 	dd if=/dev/zero of=$(disk) bs=1M count=10
@@ -28,12 +28,21 @@ buildUEFI:
 	$(parted) set 1 esp on
 
 
-runUEFI:$(disk) OVMF-pure-efi.fd
-	$(qemu) -bios OVMF-pure-efi.fd -net none
+runUEFI:buildUEFI
+	$(qemu) -hda $(disk) -bios OVMF.fd -net none
 
-compile:
+compile:clean
+	mkdir -p build
 	# compile: (flags before -o become CFLAGS in your Makefile)
-	$(CC) $(CFLAGS) -o hello.o hello.c
-	$(CC) $(CFLAGS) -o data.o path/to/gnu-efi/lib/data.c
+	$(CC) $(CFLAGS) -o build/kernel.o src/kernel.c
+	$(CC) $(CFLAGS) -o build/data.o src/data.c
 	# link: (flags before -o become LDFLAGS in your Makefile)
-	$(CC) $(LDFLAGS) -o BOOTX64.EFI hello.o data.o -lgcc
+	$(CC) $(LDFLAGS) -o build/BOOTX64.EFI build/kernel.o build/data.o -lgcc
+	dd if=/dev/zero of=build/fat.img bs=1k count=1440
+	mformat -i build/fat.img -f 1440 ::
+	mmd -i build/fat.img ::/EFI
+	mmd -i build/fat.img ::/EFI/BOOT
+	mcopy -i build/fat.img build/BOOTX64.EFI ::/EFI/BOOT
+
+compileRun:compile
+	$(qemu) -bios OVMF.fd -usb build/fat.img -net none
